Create convert.rs and implement unset logic (#31)
Move shell script conversion to a separate file, `convert.rs`.
They were previously in `main.rs`. Additionally, implement unsetting
environment variables as described in #31.
Note: until https://github.com/rust-lang/rust-clippy/pull/8271 is added
to Rust in 1.61, `#![allow(clippy::ptr_arg)]` has been added to the top
of the file.
